+++ shaku/src/guide/mod.rs
@@ -221,7 +221,9 @@
 //! parameters.
 //!
 //! Note that if you don't pass in parameters, the parameters' default values will be used. You can
-//! override the default value by annotating the property with `#[shaku(default = ...)]`.
+//! override the default value by annotating the property with `#[shaku(default = ...)]`. If the
+//! parameter should not have a default value, annotate it with `#[shaku(no_default)]`. This will
+//! cause module creation to panic if no value is provided for the parameter.
 //!
 //! ```
 //! # use shaku::{module, Component, Interface};